tegra: config: nyan-big: Add options required by Chrome OS boot
We need to match the device tree in the FIT with the U-Boot model so we
can automatically select the right device tree. Also adjust the load address
so that the device tree is not in the way when a zImage kernel tries to
extract itself.

Cover-letter:
tegra: Expand Nyan-big support
This series expands Nyan-big support:

- Enable Chrome OS EC, so that the keyboard works
- Add some extra clock and pre-kernel init required for reliable operation
- Add Chrome OS environment variables, including 'run nvboot' to allow
booting Chrome OS more easily

Still missing are audio and USB.
END
Series-changes: 2
- Drop tegra timer patch as it is not needed

Change-Id: Iddf8d061ac94ead88f97ce7ebdbdf3036bd2c8a5
Signed-off-by: Simon Glass <sjg@chromium.org>